Minor changes in login page
